refactor: Class Teacher Card name is misspelled
