+++ components/class/ClassTeacherCard.tsx
